+++ src/commands.ts
@@ -1,6 +1,9 @@
 import { checkLint } from "./commands/checkLint";
 import { navigate } from "./commands/navigate";
 import { ping } from "./commands/ping";
+import { pry } from "./commands/pry";
+import { runCurrentProject } from "./commands/runCurrentProject";
+import { runProject } from "./commands/runProject";
 import { runTestInCursor } from "./commands/runTestInCursor";
 import { runTestsInFile } from "./commands/runTestsInFile";
 import { updateDependencies } from "./commands/updateDependencies";
@@ -29,5 +32,17 @@ export const commands = [
 	{
 		command: "ouranos.checkLint",
 		handler: checkLint,
+	},
+	{
+		command: "ouranos.runCurrentProject",
+		handler: runCurrentProject,
+	},
+	{
+		command: "ouranos.runProject",
+		handler: runProject,
+	},
+	{
+		command: "ouranos.pry",
+		handler: pry,
 	},
 ];